Rebuild the homepage handler to save filled game data in case of an image error
Before, if an image error occured when uploading a game, the homepage would reload
and empty all of the game's filled fields. Now when en error occurs the title,
description and classes fields remain filled.

The game adding handler had to be removed and be rebuit into the homepage handler.