Update game field data saving to support multiple classes and fix a bug in game edit field data saving
The game field data saving in case of an error now uses dictionaries instead of
lists which improves support for saving class fields for multiple classes.
Before, if two different class titles had the same class values, after an
error both would get filled even if only one was filled in the form.

Also fixed and issue with game edit field data saving when editing a game
to have no classes. If this was done, the class fields would get filled
with the initial fields before the edit.